Call Node.__get__(parent) instead of getattr(parent, Node.name)
Node subclasses might not be attributes of the parent,
yet have references that need to be updated.

In SimBio, only one instance of Species is an attribute of the parent,
but there might be other instances
(which share the same underlying Variable)
an need to retain their stoichiometry
while replacing the underlying Variable.