Add failing test of transform with unused parameters
The compiled functions for the differential equations and transform
signature (t, y, p, out) correspond to
- t: independent
- y: variables
- p: parameters
where parameters are time-independent values.

Parameters might be functions of other constant parameters,
which can be changed without recompiling
and are resolved on Simulator.create_problem,
where (t, y, p) are created.

As the transform only receives (t, y, p) as inputs,
it must be able to be computed from these,
and is unaware of any "non-recompiling" changes in create_problem.